webapp/requirements.txt
Flask
python-docx
crewai
langchain-ollama
python-multipart